+++ docs/changelog.md
@@ -3,6 +3,9 @@ lang: en-US
 title: Changelog
 ---
 ## 2022-11-02 — New notes and updates
+- **Added** 
+  - [Software testing](/software-engineering/software-testing/) / [Software testing (ru)](/ru/software-engineering/software-testing/) 🇷🇺
+
 - **Updated**
   - [Software Engineering](/software-engineering/) / [Разработка](/ru/software-engineering/) 🇷🇺
   - [Front end](/software-engineering/front-end/) / [Front end (ru)](/ru/software-engineering/front-end/) 🇷🇺